[FEATURE] Preview of elements from BE
Some design notes on this as features many tricks and workarounds:

I would have been used a new action rather than overloading "show" in
order to:

* Avoid the "no_cache" parameter on the preview URL

But didn't go that way because of:

* We want to respect the layoutVariant configured on the "show" view.
* Using the "show" view from the "preview" action was not a choice.
* View objects do not support changing actions...

We had to stop using extbase's argument to object mapping because hidden
entities are not supported that way. We had to override "findByUid"
repository method to support that, and then get the plain integer uid
and map the object by ourserves.

That's a lot of crappy code but works.... only partially.

Workspaces preview is not supported (that seems to be fixed on 6.2).
Direct preview does not work, the workaround here is:

* Preview the page with the list plugin, versioned records will show up.
* Click on detail view, it will work.
* Won't work for hidden, versioned elements though.